Code Refactoring
Refactored TriangleFrame, etc to Image Frames to separate the jobs of
drawing Images from Triangulation and rendering Triangles.

ImageProcessor is now the main deal, it takes an Image and makes a
triangulation out of it, as well as renders the triangles on top of the
raw Image. Future Image Processing is to be done in ImageProcessor.

Bunch of test code in here to be cleaned up still.